+++ modules/sys/constants.js
@@ -70,6 +70,7 @@ export const HMCONST = {
         JAB:       1,
         BACKSTAB:  2,
         FLEEING:   3,
+        FULLPARRY: 4,
     },
 };
 
@@ -385,6 +386,23 @@ export const HMTABLES = {
             'average':      -40,
             'easy':         -80,
             'trivial':      -90,
+        },
+    },
+    statusEffects: {
+        fullparry: {
+            label: 'EFFECT.fullparry',
+            icon: 'systems/hackmaster5e/styles/icons/sword-clash.svg',
+            changes: [
+                {key: 'system.bonus.state.def',    value: '5',  mode: CONST.ACTIVE_EFFECT_MODES.ADD},
+            ],
+        },
+        sfatigue: {
+            label: 'EFFECT.sfatigue',
+            icon: 'systems/hackmaster5e/styles/icons/stoned-skull.svg',
+            changes: [
+                {key: 'system.bonus.state.def',    value: '-6',  mode: CONST.ACTIVE_EFFECT_MODES.ADD},
+                {key: 'system.bonus.state.skills', value: '-30', mode: CONST.ACTIVE_EFFECT_MODES.ADD},
+            ],
         },
     },
     'tenacity': {